Modify the memory usage calculation for the `smooth` image filter
Performing the smooth function requires copying back and forth between Gmagick
and GD, which comes at a memory cost, moreover the smooth function has a larger
memory requirement on top of that, for the convolutions. THe added factor ensures
we skip performing the filter if there is not going to be enough memory availble.