Photon: crop before resizing
Fixes #90

Resizing to a different aspect ratio requires cropping. To minimize mem usage,
the crop needs to be performed before resizing.

The aggressive crop strategy can introduce minor differences in the border
pixels (due to having less data available for interpolation), but this should
not be an issue as they are expected to be visually unnoticeable.