photon: Rework webp conversion
Refactors the webp generation code so that these are no repeated bits in the
jpg and png functions, and all actions are explicit.

Attempts to generate the webp file from the image data in RAM, not the already
encoded fallback file. Since the backend libs do not support webp directly, an
intermediate lossless file with as little compression as possible (for speed)
is used for the conversion. This addresses and fixes #88.

Also prevents jpegs from being encoded as lossless webps, which does not
increase quality, but often results in larger files.

A simple test for this change is requesting a png and jpeg file, each with and
without the `Accept: image/webp` header. All 4 files should be encoded
correctly, and have the expected quality. See also #88 for more info on how
webps could have lower than expected quality, due to them being encoded from a
lossy file.